requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
torch==2.8.0
torchvision==0.23.0
ultralytics==8.3.189
requests==2.32.5
python-dotenv==1.0.0
